test(functional): split shared agent fixture
